Optmisation using a simple cache table.
