+++ webapp/src/main/java/org/apache/atlas/web/rest/EntityREST.java
@@ -164,74 +164,7 @@ public class EntityREST {
             if (AtlasPerfTracer.isPerfTraceEnabled(PERF_LOG)) {
                 perf = AtlasPerfTracer.getPerfTracer(PERF_LOG, "EntityREST.evaluatePolicies()");
             }
-
-            for (int i = 0; i < entities.size(); i++) {
-
-                String action = entities.get(i).getAction();
-
-                if (action == null) {
-                    throw new AtlasBaseException(BAD_REQUEST, "action is null");
-                }
-
-                if (ENTITY_READ.name().equals(action) || ENTITY_CREATE.name().equals(action) || ENTITY_UPDATE.name().equals(action)
-                        || ENTITY_DELETE.name().equals(action) || ENTITY_UPDATE_BUSINESS_METADATA.name().equals(action)) {
-
-                    try {
-                        AtlasEntityHeader entityHeader = getAtlasEntityHeader(entities.get(i).getEntityGuid(), entities.get(i).getEntityId(),entities.get(i).getTypeName());
-
-                        AtlasEntityAccessRequest.AtlasEntityAccessRequestBuilder requestBuilder = new AtlasEntityAccessRequest.AtlasEntityAccessRequestBuilder(typeRegistry, AtlasPrivilege.valueOf(entities.get(i).getAction()), entityHeader);
-                        if (entities.get(i).getBusinessMetadata() != null) {
-                            requestBuilder.setBusinessMetadata(entities.get(i).getBusinessMetadata());
-                        }
-
-                        AtlasEntityAccessRequest entityAccessRequest = requestBuilder.build();
-
-                        AtlasAuthorizationUtils.verifyAccess(entityAccessRequest, entities.get(i).getAction() + "guid=" + entities.get(i).getEntityGuid());
-                        response.add(new AtlasEvaluatePolicyResponse(entities.get(i).getTypeName(), entities.get(i).getEntityGuid(), entities.get(i).getAction(), entities.get(i).getEntityId(), true, null , entities.get(i).getBusinessMetadata()));
-                    } catch (AtlasBaseException e) {
-                        AtlasErrorCode code = e.getAtlasErrorCode();
-                        String errorCode = code.getErrorCode();
-                        response.add(new AtlasEvaluatePolicyResponse(entities.get(i).getTypeName(), entities.get(i).getEntityGuid(), entities.get(i).getAction(), entities.get(i).getEntityId(), false, errorCode, entities.get(i).getBusinessMetadata()));
-                    }
-
-                } else if (ENTITY_REMOVE_CLASSIFICATION.name().equals(action) || ENTITY_ADD_CLASSIFICATION.name().equals(action) || ENTITY_UPDATE_CLASSIFICATION.name().equals(action)) {
-
-                    if (entities.get(i).getClassification() == null) {
-                        throw new AtlasBaseException(BAD_REQUEST, "classification needed for " + action + " authorization");
-                    }
-                    try {
-                        AtlasEntityHeader entityHeader = getAtlasEntityHeader(entities.get(i).getEntityGuid(), entities.get(i).getEntityId(),entities.get(i).getTypeName());
-
-                        AtlasAuthorizationUtils.verifyAccess(new AtlasEntityAccessRequest(typeRegistry, AtlasPrivilege.valueOf(entities.get(i).getAction()), entityHeader, new AtlasClassification(entities.get(i).getClassification())));
-                        response.add(new AtlasEvaluatePolicyResponse(entities.get(i).getTypeName(), entities.get(i).getEntityGuid(), entities.get(i).getAction(), entities.get(i).getEntityId(), entities.get(i).getClassification(), true, null));
-
-                    } catch (AtlasBaseException e) {
-                        AtlasErrorCode code = e.getAtlasErrorCode();
-                        String errorCode = code.getErrorCode();
-                        response.add(new AtlasEvaluatePolicyResponse(entities.get(i).getTypeName(), entities.get(i).getEntityGuid(), entities.get(i).getAction(), entities.get(i).getEntityId(), entities.get(i).getClassification(), false, errorCode));
-                    }
-
-                }    else if (RELATIONSHIP_ADD.name().equals(action) || RELATIONSHIP_REMOVE.name().equals(action) || RELATIONSHIP_UPDATE.name().equals(action)) {
-
-                    if (entities.get(i).getRelationShipTypeName() == null) {
-                        throw new AtlasBaseException(BAD_REQUEST, "RelationShip TypeName needed for " + action + " authorization");
-                    }
-
-                    try {
-                        AtlasEntityHeader end1Entity = getAtlasEntityHeader(entities.get(i).getEntityGuidEnd1(), entities.get(i).getEntityIdEnd1(), entities.get(i).getEntityTypeEnd1());
-
-                        AtlasEntityHeader end2Entity = getAtlasEntityHeader(entities.get(i).getEntityGuidEnd2(), entities.get(i).getEntityIdEnd2(), entities.get(i).getEntityTypeEnd2());
-
-                        AtlasAuthorizationUtils.verifyAccess(new AtlasRelationshipAccessRequest(typeRegistry, AtlasPrivilege.valueOf(action), entities.get(i).getRelationShipTypeName(), end1Entity, end2Entity));
-                        response.add(new AtlasEvaluatePolicyResponse(action, entities.get(i).getRelationShipTypeName(), entities.get(i).getEntityTypeEnd1(), entities.get(i).getEntityGuidEnd1(), entities.get(i).getEntityIdEnd1(), entities.get(i).getEntityTypeEnd2(), entities.get(i).getEntityGuidEnd2(), entities.get(i).getEntityIdEnd2(), true, null));
-                    } catch (AtlasBaseException e) {
-                        AtlasErrorCode code = e.getAtlasErrorCode();
-                        String errorCode = code.getErrorCode();
-                        response.add(new AtlasEvaluatePolicyResponse(action, entities.get(i).getRelationShipTypeName(), entities.get(i).getEntityTypeEnd1(), entities.get(i).getEntityGuidEnd1(), entities.get(i).getEntityIdEnd1(), entities.get(i).getEntityTypeEnd2(), entities.get(i).getEntityGuidEnd2(), entities.get(i).getEntityIdEnd2(), false, errorCode));
-                    }
-                }
-
-            }
+            response = entitiesStore.evaluatePolicies(entities);
         } finally {
             AtlasPerfTracer.log(perf);
         }
@@ -261,28 +194,6 @@ public class EntityREST {
             AtlasPerfTracer.log(perf);
         }
         return ret;
-    }
-
-    private AtlasEntityHeader getAtlasEntityHeader(String entityGuid, String entityId, String entityType) throws AtlasBaseException {
-        AtlasEntityHeader entityHeader = null;
-
-        if (StringUtils.isNotEmpty(entityGuid)) {
-            AtlasEntityWithExtInfo ret = entitiesStore.getByIdWithoutAuthorization(entityGuid);
-            entityHeader = new AtlasEntityHeader(ret.getEntity());
-        } else if (StringUtils.isNotEmpty(entityId) && StringUtils.isNotEmpty(entityType)) {
-            try {
-                entityHeader = entitiesStore.getAtlasEntityHeaderWithoutAuthorization(null, entityId, entityType);
-            } catch (AtlasBaseException abe) {
-                if (abe.getAtlasErrorCode() == AtlasErrorCode.INSTANCE_BY_UNIQUE_ATTRIBUTE_NOT_FOUND) {
-                    Map<String, Object> attributes = new HashMap<>();
-                    attributes.put(QUALIFIED_NAME, entityId);
-                    entityHeader = new AtlasEntityHeader(entityType, attributes);
-                }
-            }
-        } else {
-            throw new AtlasBaseException(BAD_REQUEST, "requires entityGuid or typeName and qualifiedName for entity authorization");
-        }
-        return entityHeader;
     }
